[eclipse-iceoryx#956] use unsigned datatype to fix tests

+++ examples/cxx/health_monitoring/src/publisher_2.cpp
@@ -36,7 +36,7 @@ auto main() -> int {
                         .default_event_id(iox::into<EventId>(PubSubEvent::SentSample))
                         .create()
                         .expect("");
-    auto counter = 1000000; // NOLINT, magic number is fine in an example
+    auto counter = 1000000U; // NOLINT, magic number is fine in an example
 
     auto waitset = WaitSetBuilder().create<ServiceType::Ipc>().expect("");
 
@@ -47,7 +47,7 @@ auto main() -> int {
     waitset
         .wait_and_process([&](auto) {
             std::cout << service_name.to_string().c_str() << ": Send sample " << counter << " ..." << std::endl;
-            publisher.send_copy(static_cast<uint64_t>(counter)).expect("");
+            publisher.send_copy(counter).expect("");
             notifier.notify().expect("");
             counter += 1;
             return CallbackProgression::Continue;
